Quang | Database migration with data.sql and match with schema.sql

+++ backend/src/main/java/com/example/demo/booking/Booking.java
@@ -20,7 +20,7 @@ public class Booking {
     @GeneratedValue(strategy = GenerationType.AUTO, generator = "my_seq")
     @SequenceGenerator(name = "my_seq", sequenceName = "booking_seq", allocationSize = 1)
     @Column(name = "booking_id")
-    private Long bookingID;
+    private Integer bookingID;
 
     @Column(name = "booking_date")
     private Date bookingDate;
